Fix Firefox hang when Influx is down
Removes the user agent check for Firefox that used to send the final
request asynchronous. If Influx was down, the final request would hang
the UI: https://bugzilla.mozilla.org/show_bug.cgi?id=1103323.

We should be getting enough metrics from Firefox during runtime, and
using the Beacon API. We might want to send the data incrementally at
some point, so we can completely do away with the final request if it
fails.